Add editorUtils
