+++ twake/frontend/src/app/components/RichTextEditor/Editor.tsx
@@ -11,6 +11,7 @@ import MentionSuggestion from "./components/mentions/MentionSuggestion";
 import ChannelSuggestion from "./components/channel/ChannelSuggestion";
 import "./Editor.scss";
 import CommandSuggestion from "./components/commands/CommandSuggestion";
+import { getCaretCoordinates, getInsertRange, getTrigger } from "./EditorUtils";
 
 const { isSoftNewlineEvent } = KeyBindingUtil;
 
@@ -46,6 +47,7 @@ export type EditorSuggestionPlugin<SuggestionType> = {
   decorator: DraftDecorator;
   trigger: string | RegExp;
   resourceType: string;
+  getTextDisplay?: (item: SuggestionType) => string;
 };
 
 type EditorViewState = {
@@ -402,100 +404,6 @@ export class EditorView extends React.Component<EditorProps, EditorViewState> {
     </div>
   }
 }
-
-function getSelectionRange() {
-  const selection = window.getSelection()
-  if (selection?.rangeCount === 0) {
-    return null
-  }
-  return selection?.getRangeAt(0)
-}
-
-function getTrigger(trigger: string | RegExp) {
-  if (typeof trigger === "string") {
-    return getTriggerRange(trigger);
-  } else if (trigger instanceof RegExp) {
-    return getTriggerMatchRange(trigger);
-  }
-}
-
-function getTriggerMatchRange(regexp: RegExp) {
-  const range = getSelectionRange();
-  if (!range) {
-    return null;
-  }
-  const text  = range && range?.startContainer?.textContent?.substring(0, range.startOffset)
-  if (!text || /\s+$/.test(text)) {
-    return null;
-  }
-
-  const start = text.match(regexp);
-
-  if (!start) {
-    return;
-  }
-
-  const end = range.startOffset
-  return {
-    end,
-    start,
-    text: start[1],
-  }
-}
-
-function getTriggerRange(term: string) {
-  const range = getSelectionRange();
-  if (!range) {
-    return null;
-  }
-  const text  = range && range?.startContainer?.textContent?.substring(0, range.startOffset)
-  if (!text || /\s+$/.test(text))
-    return null
-
-  const start = text.lastIndexOf(term)
-  if (start === -1) {
-    return null
-  }
-
-  const end = range.startOffset
-  return {
-    end,
-    start,
-    text: text.substring(start),
-  }
-}
-
-function getInsertRange(editorState: EditorState, firstCharacter: string): { start: number, end: number } {
-  const selection = editorState.getSelection();
-  const content = editorState.getCurrentContent();
-  const anchorKey = selection.getAnchorKey();
-  const end = selection.getAnchorOffset();
-  const block = content.getBlockForKey(anchorKey);
-  const text = block.getText();
-  const start = text.substring(0, end).lastIndexOf(firstCharacter);
-
-  return {
-    start,
-    end,
-  };
-}
-
-const currentCoordinates: CaretCoordinates = {
-  x: 0,
-  y: 0,
-};
-
-function getCaretCoordinates(): CaretCoordinates {
-  const range = getSelectionRange();
-  if (range) {
-    const { left: x, top: y }  = range.getBoundingClientRect();
-    Object.assign(currentCoordinates, { x, y });
-  }
-  return currentCoordinates;
-}
-
-
-// TODO: Extract API
 
 const addMention = (editorState: EditorState, mention: MentionSuggestionType, prefix: string): EditorState => {
   const { start, end } = getInsertRange(editorState, prefix)
